updated animation performance for virtual thought
