+++ src/components/VirtualThought.tsx
@@ -19,6 +19,8 @@ import noteValue from '../util/noteValue'
 import DropChild from './DropChild'
 import DropUncle from './DropUncle'
 import Subthought from './Subthought'
+import useChangeRef from '../hooks/useChangeRef';
+import { css } from '../../styled-system/css';
 
 /** A resize handler that should be called whenever a thought's height has changed. */
 export type OnResize = (args: {
@@ -96,6 +98,7 @@ const VirtualThought = ({
   const fontSize = useSelector(state => state.fontSize)
   const note = useSelector(state => noteValue(state, thought.id))
   const ref = useRef<HTMLDivElement>(null)
+  const autofocusChanged = useChangeRef(autofocus)
 
   /***************************
    * VirtualThought properties
@@ -201,6 +204,16 @@ const VirtualThought = ({
     [crossContextualKey, onResize, thought.id],
   )
 
+  // When autofocus changes, use a slow (750ms) ease-out to provide a gentle transition to non-focal thoughts.
+  // If autofocus has not changed, it means that the thought is being rendered for the first time, such as the children of a thought that was just expanded. In this case, match the tree-node top animation (150ms) to ensure that the newly rendered thoughts fade in to fill the space that is being opened up from the next uncle animating down.
+  // Note that ease-in is used in contrast to the tree-node's ease-out. This gives a little more time for the next uncle to animate down and clear space before the newly rendered thought fades in. Otherwise they overlap too much during the transition.
+  const opacity = autofocus === 'show' ? '1' : autofocus === 'dim' ? '0.5' : '0'
+  useEffect(() => {
+    if (!ref.current) return
+    // start opacity at 0 and set to actual opacity in useEffect
+    ref.current.style.opacity = opacity
+  })
+
   // Short circuit if thought has already been removed.
   // This can occur in a re-render even when thought is defined in the parent component.
   if (!thought) return null
@@ -208,6 +221,21 @@ const VirtualThought = ({
   return (
     <div
       ref={ref}
+      className={css({
+        // Start opacity at 0 and set to actual opacity in useEffect.
+        // Do not fade in empty thoughts. An instant snap in feels better here.
+        // opacity creates a new stacking context, so it must only be applied to Thought, not to the outer VirtualThought which contains DropChild. Otherwise subsequent DropChild will be obscured.
+        opacity: thought.value === '' ? opacity : '0',
+        transition: autofocusChanged
+          ? `opacity {durations.layoutSlowShiftDuration} ease-out`
+          : `opacity {durations.layoutNodeAnimationDuration} ease-in`,
+        pointerEvents: !isVisible ? 'none' : undefined,
+        // Safari has a known issue with subpixel calculations, especially during animations and with SVGs.
+        // This caused the thought to jerk slightly to the left at the end of the horizontal shift animation.
+        // By setting "will-change: transform;", we hint to the browser that the transform property will change in the future,
+        // allowing the browser to optimize the animation.
+        willChange: 'opacity',
+      })}
       style={{
         // Fix the height of the container to the last measured height to ensure that there is no layout shift when the Thought is removed from the DOM.
         // Must include DropChild, or it will shift when the cursor moves.
